refactor: rename pricing modules to model_pricing
