+++ crates/core/tests/unit/observability/otel_tests.rs
@@ -13,7 +13,7 @@ use crate::api::runtime::global_context;
 use crate::api::scope::ScopeType;
 use crate::api::scope::{event, pop_scope, push_scope};
 use crate::api::tool::ToolAttributes;
-use crate::codec::pricing::pricing_test_mutex;
+use crate::codec::model_pricing::pricing_test_mutex;
 use crate::codec::response::{
     AnnotatedLlmResponse, CostEstimate, CostSource, PricingCatalog, PricingResolver, Usage,
     reset_active_pricing_resolver, set_active_pricing_resolver,
