+++ crates/wasm/tests-js/index_loader_tests.mjs
@@ -18,7 +18,7 @@ test('WebAssembly generated package exposes the expected package metadata', () =
   assert.equal(packageJson.exports['./plugin'].default, './plugin.js');
   assert.equal(packageJson.exports['./adaptive'].default, './adaptive.js');
   assert.equal(packageJson.exports['./pii_redaction'].default, './pii_redaction.js');
-  assert.equal(packageJson.exports['./pricing'].default, './pricing.js');
+  assert.equal(packageJson.exports['./model_pricing'].default, './model_pricing.js');
   assert.equal(packageJson.exports['./typed.js'].default, './typed.js');
   assert.equal(typeof wasm.ScopeType.Agent, 'number');
   assert.equal(wasm.ScopeType.Agent, 0);
@@ -36,8 +36,8 @@ test('WebAssembly generated package includes the expected wrapper files', () => 
     'adaptive.d.ts',
     'pii_redaction.js',
     'pii_redaction.d.ts',
-    'pricing.js',
-    'pricing.d.ts',
+    'model_pricing.js',
+    'model_pricing.d.ts',
   ];
 
   for (const fileName of expectedFiles) {
