+++ crates/wasm/wrappers/esm/model_pricing.js
@@ -1,30 +1,28 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-'use strict';
+import * as plugin from './plugin.js';
 
-const plugin = require('./plugin.js');
-
-const PRICING_PLUGIN_KIND = 'pricing';
+export const PRICING_PLUGIN_KIND = 'pricing';
 
 /**
- * Create a default pricing component config.
+ * Create a default model pricing component config.
  *
- * @returns {object} The minimal pricing config with no sources.
+ * @returns {object} The minimal model pricing config with no sources.
  */
-function defaultConfig() {
+export function defaultConfig() {
   return {
     sources: [],
   };
 }
 
 /**
- * Create per-token pricing rates with defaults applied.
+ * Create per-token model pricing rates with defaults applied.
  *
  * @param {object} [config={}] - Partial token-rate fields to override.
- * @returns {object} A normalized token pricing rates object.
+ * @returns {object} A normalized token model pricing rates object.
  */
-function tokenRates(config = {}) {
+export function tokenRates(config = {}) {
   return {
     input_per_million: 0,
     output_per_million: 0,
@@ -38,7 +36,7 @@ function tokenRates(config = {}) {
  * @param {object} [config={}] - Partial prompt-cache settings to override.
  * @returns {object} A normalized prompt-cache config object.
  */
-function promptCache(config = {}) {
+export function promptCache(config = {}) {
   return {
     read_accounting: 'included_in_prompt_tokens',
     ...config,
@@ -52,7 +50,7 @@ function promptCache(config = {}) {
  * @param {object} [config={}] - Optional threshold fields to include.
  * @returns {object} A normalized token rate tier object.
  */
-function tokenRateTier(rates, config = {}) {
+export function tokenRateTier(rates, config = {}) {
   return {
     ...config,
     rates,
@@ -66,7 +64,7 @@ function tokenRateTier(rates, config = {}) {
  * @param {object} [config={}] - Optional schedule fields to override.
  * @returns {object} A normalized rate schedule object.
  */
-function promptTokenThresholdRateSchedule(tiers = [], config = {}) {
+export function promptTokenThresholdRateSchedule(tiers = [], config = {}) {
   return {
     type: 'prompt_token_threshold',
     applies_to: 'full_request',
@@ -76,30 +74,29 @@ function promptTokenThresholdRateSchedule(tiers = [], config = {}) {
 }
 
 /**
- * Create one pricing catalog entry with defaults applied.
+ * Create one model pricing catalog entry with defaults applied.
  *
  * @param {object} config - Required and optional model pricing fields.
  * @returns {object} A normalized model pricing entry.
  */
-function catalogEntry(config) {
-  const { prompt_cache: promptCacheConfig, ...entryConfig } = config;
+export function catalogEntry(config) {
   return {
     aliases: [],
     currency: 'USD',
     unit: 'per_token',
-    ...entryConfig,
-    prompt_cache: promptCache(promptCacheConfig),
+    prompt_cache: promptCache(),
+    ...config,
   };
 }
 
 /**
- * Create an inline pricing catalog payload.
+ * Create an inline model pricing catalog payload.
  *
  * @param {object[]} [entries=[]] - Pricing catalog entries.
  * @param {object} [config={}] - Optional catalog fields to override.
- * @returns {object} A normalized pricing catalog object.
+ * @returns {object} A normalized model pricing catalog object.
  */
-function inlineCatalog(entries = [], config = {}) {
+export function inlineCatalog(entries = [], config = {}) {
   return {
     version: 1,
     entries,
@@ -108,12 +105,12 @@ function inlineCatalog(entries = [], config = {}) {
 }
 
 /**
- * Create an inline pricing source.
+ * Create an inline model pricing source.
  *
  * @param {object} catalog - Pricing catalog payload.
  * @returns {object} A normalized inline source config.
  */
-function inlineSource(catalog) {
+export function inlineSource(catalog) {
   return {
     type: 'inline',
     catalog,
@@ -121,12 +118,12 @@ function inlineSource(catalog) {
 }
 
 /**
- * Create a file-backed pricing source.
+ * Create a file-backed model pricing source.
  *
  * @param {string} path - JSON catalog file path.
  * @returns {object} A normalized file source config.
  */
-function fileSource(path) {
+export function fileSource(path) {
   return {
     type: 'file',
     path,
@@ -134,55 +131,39 @@ function fileSource(path) {
 }
 
 /**
- * Create a pricing config from ordered sources.
+ * Create a model pricing config from ordered sources.
  *
  * @param {object[]} [sources=[]] - Pricing sources in precedence order.
- * @returns {object} A normalized pricing config object.
+ * @returns {object} A normalized model pricing config object.
  */
-function pricingConfig(sources = []) {
+export function pricingConfig(sources = []) {
   return {
     sources,
   };
 }
 
 /**
- * Wrap pricing config as a top-level plugin component.
+ * Wrap model pricing config as a top-level plugin component.
  *
- * @param {object} config - Pricing component configuration document.
+ * @param {object} config - Model pricing component configuration document.
  * @param {{ enabled?: boolean }} [options={}] - Optional component-level flags.
- * @returns {object} A plugin component spec for the pricing plugin.
+ * @returns {object} A plugin component spec for the model pricing plugin.
  */
-function ComponentSpec(config, { enabled = true } = {}) {
+export function ComponentSpec(config, { enabled = true } = {}) {
   return plugin.ComponentSpec(PRICING_PLUGIN_KIND, config, {
     enabled,
   });
 }
 
 /**
- * Validate a pricing config document without activating it.
+ * Validate a model pricing config document without activating it.
  *
- * @param {object} config - Pricing component configuration document.
+ * @param {object} config - Model pricing component configuration document.
  * @returns {object} A structured validation report with diagnostics.
  */
-function validateConfig(config) {
+export function validateConfig(config) {
   return plugin.validate({
     version: 1,
     components: [ComponentSpec(config)],
   });
 }
-
-module.exports = {
-  PRICING_PLUGIN_KIND,
-  defaultConfig,
-  tokenRates,
-  promptCache,
-  tokenRateTier,
-  promptTokenThresholdRateSchedule,
-  catalogEntry,
-  inlineCatalog,
-  inlineSource,
-  fileSource,
-  pricingConfig,
-  ComponentSpec,
-  validateConfig,
-};
