+++ go/nemo_relay/model_pricing/model_pricing.go
@@ -1,23 +1,23 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package pricing
+package model_pricing
 
 import nemo_relay "github.com/NVIDIA/NeMo-Relay/go/nemo_relay"
 
-// Config is the canonical pricing config document.
+// Config is the canonical model pricing config document.
 type Config = nemo_relay.PricingConfig
 
-// SourceConfig is implemented by pricing source config structs.
+// SourceConfig is implemented by model pricing source config structs.
 type SourceConfig = nemo_relay.PricingSourceConfigurer
 
-// InlineSourceConfig embeds a pricing catalog directly in plugin config.
+// InlineSourceConfig embeds a model pricing catalog directly in plugin config.
 type InlineSourceConfig = nemo_relay.PricingInlineSourceConfig
 
-// FileSourceConfig loads a pricing catalog from a JSON file path.
+// FileSourceConfig loads a model pricing catalog from a JSON file path.
 type FileSourceConfig = nemo_relay.PricingFileSourceConfig
 
-// Catalog is an inline pricing catalog payload.
+// Catalog is an inline model pricing catalog payload.
 type Catalog = nemo_relay.PricingCatalog
 
 // ModelPricing is one model pricing catalog entry.
@@ -26,7 +26,7 @@ type ModelPricing = nemo_relay.ModelPricing
 // TokenRates expresses model rates per one million tokens.
 type TokenRates = nemo_relay.TokenPricingRates
 
-// PromptCacheConfig configures cache-read token accounting for a pricing entry.
+// PromptCacheConfig configures cache-read token accounting for a model pricing entry.
 type PromptCacheConfig = nemo_relay.PromptCachePricing
 
 // TokenRateTier is one prompt-token threshold tier.
@@ -35,16 +35,16 @@ type TokenRateTier = nemo_relay.TokenRateTier
 // PromptTokenThresholdRateSchedule selects token rates by prompt token thresholds.
 type PromptTokenThresholdRateSchedule = nemo_relay.PromptTokenThresholdRateSchedule
 
-// ComponentSpec wraps pricing config as a top-level pricing component.
+// ComponentSpec wraps model pricing config as a top-level model pricing component.
 type ComponentSpec = nemo_relay.PricingComponentSpec
 
 // ConfigReport is the validation or activation report for a plugin config.
 type ConfigReport = nemo_relay.ConfigReport
 
-// PluginKind is the top-level plugin kind used by the pricing component.
+// PluginKind is the top-level plugin kind used by the model pricing component.
 const PluginKind = nemo_relay.PricingPluginKind
 
-// Pricing unit constants accepted by pricing catalog entries.
+// Pricing unit constants accepted by model pricing catalog entries.
 const (
 	UnitPerToken   = nemo_relay.PricingUnitPerToken
 	UnitPerRequest = nemo_relay.PricingUnitPerRequest
@@ -52,28 +52,28 @@ const (
 	UnitGPUHour    = nemo_relay.PricingUnitGPUHour
 )
 
-// Prompt-cache accounting constants accepted by pricing catalog entries.
+// Prompt-cache accounting constants accepted by model pricing catalog entries.
 const (
 	CacheReadIncludedInPromptTokens = nemo_relay.CacheReadAccountingIncludedInPromptTokens
 	CacheReadSeparate               = nemo_relay.CacheReadAccountingSeparate
 )
 
-// NewConfig returns an empty pricing config.
+// NewConfig returns an empty model pricing config.
 func NewConfig() Config {
 	return nemo_relay.NewPricingConfig()
 }
 
-// NewInlineSource returns an inline pricing catalog source.
+// NewInlineSource returns an inline model pricing catalog source.
 func NewInlineSource(catalog Catalog) InlineSourceConfig {
 	return nemo_relay.NewPricingInlineSource(catalog)
 }
 
-// NewFileSource returns a file-backed pricing catalog source.
+// NewFileSource returns a file-backed model pricing catalog source.
 func NewFileSource(path string) FileSourceConfig {
 	return nemo_relay.NewPricingFileSource(path)
 }
 
-// NewCatalog returns an inline pricing catalog with version 1.
+// NewCatalog returns an inline model pricing catalog with version 1.
 func NewCatalog(entries ...ModelPricing) Catalog {
 	return nemo_relay.NewPricingCatalog(entries...)
 }
@@ -103,17 +103,17 @@ func NewPromptTokenThresholdRateSchedule(tiers ...TokenRateTier) PromptTokenThre
 	return nemo_relay.NewPromptTokenThresholdRateSchedule(tiers...)
 }
 
-// NewComponentSpec wraps pricing config as an enabled top-level component.
+// NewComponentSpec wraps model pricing config as an enabled top-level component.
 func NewComponentSpec(config Config) ComponentSpec {
 	return nemo_relay.NewPricingComponentSpec(config)
 }
 
-// Component converts pricing config directly into the shared plugin shape.
+// Component converts model pricing config directly into the shared plugin shape.
 func Component(config Config) nemo_relay.PluginComponentSpec {
 	return nemo_relay.PricingComponent(config)
 }
 
-// ValidateConfig validates a pricing config document without activating it.
+// ValidateConfig validates a model pricing config document without activating it.
 func ValidateConfig(config Config) (ConfigReport, error) {
 	return nemo_relay.ValidatePricingConfig(config)
 }
