+++ python/nemo_relay/__init__.py
@@ -17,7 +17,7 @@ The main entry points are:
 - ``nemo_relay.adaptive`` for adaptive component configuration helpers
 - ``nemo_relay.observability`` for observability component configuration helpers
 - ``nemo_relay.pii_redaction`` for PII redaction component configuration helpers
-- ``nemo_relay.pricing`` for model pricing component configuration helpers
+- ``nemo_relay.model_pricing`` for model pricing component configuration helpers
 
 Top-level exports also include:
 
@@ -191,10 +191,10 @@ from nemo_relay import (  # noqa: E402
     guardrails,
     intercepts,
     llm,
+    model_pricing,
     observability,
     pii_redaction,
     plugin,
-    pricing,
     scope,
     scope_local,
     subscribers,
@@ -438,7 +438,7 @@ __all__ = [
     "adaptive",
     "observability",
     "pii_redaction",
-    "pricing",
+    "model_pricing",
     # Scope stack isolation
     "ScopeStack",
     "create_scope_stack",
