+++ python/nemo_relay/__init__.pyi
@@ -30,10 +30,10 @@ from nemo_relay import codecs as codecs
 from nemo_relay import guardrails as guardrails
 from nemo_relay import intercepts as intercepts
 from nemo_relay import llm as llm
+from nemo_relay import model_pricing as model_pricing
 from nemo_relay import observability as observability
 from nemo_relay import pii_redaction as pii_redaction
 from nemo_relay import plugin as plugin
-from nemo_relay import pricing as pricing
 from nemo_relay import scope as scope
 from nemo_relay import scope_local as scope_local
 from nemo_relay import subscribers as subscribers
